python-backend/requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# AI/ML
openai==1.10.0
numpy==1.26.3

# Background Tasks
celery==5.3.6
redis==5.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1
pytz==2023.3
email-validator==2.1.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14

